meta: improve build system
use pkg-config!
the main concern at this point is whether or not adding a "--static" is
safe.
if you don't add "--static", there's no guarantee of static linking.
however, if you do add "--static" but not all of the libraries in
question are available in static form, it will fail to build.
the likely correct way to approach this is to use an env var.
furthermore, I'm not entirely sure as to how/whether the link flags get
passed on to upstream packages.