patch against getcbytes bug, to be tested on linux
I didn't catch this in development because macos helpfully zeroes out
all malloc pages on allocation, so my buffers *were* \0-terminated.
Getcbytes is called in a few spots, namely in setopt, and 4 mime
functions.
We only ever call setopt using the partial `pt`, so we can just `comp`
it with `freeze`.
As for mime, we insert a `freeze` in each of the 4 handlers.
Either way, also backport the new getcbytes, which still has the failure
condition, but only under very specific circumstances.